refactor: add Image folder, downloadable prop
